pyproject.toml
[project]
name = "clean-fastapi"
version = "0.1.0"
description = "Clean Architecture boilerplate for FastAPI apps"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "argon2-cffi>=25.1.0",
    "babel>=2.17.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "msgspec>=0.19.0",
    "psycopg[binary]>=3.2.9",
    "pyjwt>=2.10.1",
    "pytest-mock>=3.14.1",
    "python-decouple>=3.8",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "faker>=37.4.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.8.0",
    "time-machine>=2.16.0",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
strict = true
disallow_subclassing_any = false
disallow_untyped_decorators = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = ["no-untyped-def"]

[tool.coverage.run]
omit = [
    "tests/*",
    "*/tests/*",
    "*/**/*.html",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abc.abstractmethod",
]
